pyproject.toml
[project]
name = "discord-welcome-bot"
version = "0.1.0"
description = "Discord bot that welcomes new users"
requires-python = ">=3.11,<3.13"

dependencies = [
    "discord.py",
    "python-dotenv"
]